backend/requirements.txt
#pycrypto>=2.6
psycopg2-binary>=2.9
itsdangerous>=2.1
sqlalchemy>=2.0
authlib>=1.3
openpyxl>=3.1
pillow>=10.0

flask>=3.0
click>=8.0
flask-sqlalchemy>=3.0
flask-migrate>=4.0
flask_mail>=0.9

pytest>=8.0
coverage>=7.4
